requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
numba>=0.56.0
MDAnalysis>=2.0.0
biopython>=1.79
seaborn>=0.11.0
pytest>=6.0.0
py3Dmol>=2.0.0  # 可视化
